O método confirmRide agora recebe a distancia do percurso na medida correta (quilometros).

+++ app/src/main/java/com/phoenix/rideapp/feature_ride/data/repository/RideApiRepositoryImpl.kt
@@ -79,7 +79,7 @@ class RideApiRepositoryImpl @Inject constructor (
     override suspend fun confirmRide(
         customerId: String,
         destination: String,
-        distance: Int,
+        distance: Double,
         driver: Driver,
         duration: String,
         origin: String,
