+++ app/src/main/java/com/phoenix/rideapp/feature_ride/domain/model/ride_api/RideApiRepository.kt
@@ -38,7 +38,7 @@ interface RideApiRepository {
     suspend fun confirmRide(
         customerId: String,
         destination: String,
-        distance: Int,
+        distance: Double,
         driver: Driver,
         duration: String,
         origin: String,
